updating archvis for render task data sources to renderers
