--- plugins/archvis_gl/src/FEMRenderTaskDataSource.h
@@ -1,65 +1,0 @@
-/**
- * MegaMol
- * Copyright (c) 2019, MegaMol Dev Team
- * All rights reserved.
- */
-
-#pragma once
-
-#include "mesh_gl/AbstractGPURenderTaskDataSource.h"
-
-namespace megamol::archvis_gl {
-
-class FEMRenderTaskDataSource : public mesh_gl::AbstractGPURenderTaskDataSource {
-public:
-    /**
-     * Answer the name of this module.
-     *
-     * @return The name of this module.
-     */
-    static const char* ClassName(void) {
-        return "FEMGPURenderTasksDataSource";
-    }
-
-    /**
-     * Answer a human readable description of this module.
-     *
-     * @return A human readable description of this module.
-     */
-    static const char* Description(void) {
-        return "Data source for loading render tasks based on FEM data";
-    }
-
-    /**
-     * Answers whether this module is available on the current system.
-     *
-     * @return 'true' if the module is available, 'false' otherwise.
-     */
-    static bool IsAvailable(void) {
-        return true;
-    }
-
-
-    FEMRenderTaskDataSource();
-    ~FEMRenderTaskDataSource();
-
-protected:
-    /**
-     * Implementation of 'Create'.
-     *
-     * @return 'true' on success, 'false' otherwise.
-     */
-    virtual bool create(void);
-
-    virtual bool getDataCallback(core::Call& caller);
-
-private:
-    uint32_t m_version;
-
-    megamol::core::CallerSlot m_fem_callerSlot;
-
-    /** In-place material collection (initialized with gltf btf) */
-    std::shared_ptr<mesh_gl::GPUMaterialCollection> m_material_collection;
-};
-
-} // namespace megamol::archvis_gl
