--- plugins/archvis_gl/src/MSMRenderTaskDataSource.h
@@ -1,59 +1,0 @@
-/**
- * MegaMol
- * Copyright (c) 2019, MegaMol Dev Team
- * All rights reserved.
- */
-
-#pragma once
-
-#include "mesh_gl/AbstractGPURenderTaskDataSource.h"
-
-namespace megamol::archvis_gl {
-
-class MSMRenderTaskDataSource : public mesh_gl::AbstractGPURenderTaskDataSource {
-public:
-    /**
-     * Answer the name of this module.
-     *
-     * @return The name of this module.
-     */
-    static const char* ClassName(void) {
-        return "MSMRenderTaskDataSource";
-    }
-
-    /**
-     * Answer a human readable description of this module.
-     *
-     * @return A human readable description of this module.
-     */
-    static const char* Description(void) {
-        return "Data source for creating RenderTasks for given MSM input data.";
-    }
-
-    /**
-     * Answers whether this module is available on the current system.
-     *
-     * @return 'true' if the module is available, 'false' otherwise.
-     */
-    static bool IsAvailable(void) {
-        return true;
-    }
-
-    MSMRenderTaskDataSource();
-    ~MSMRenderTaskDataSource();
-
-protected:
-    virtual bool getDataCallback(core::Call& caller);
-
-    virtual bool getMetaDataCallback(core::Call& caller) override;
-
-private:
-    megamol::core::CallerSlot m_MSM_callerSlot;
-
-    /** In-place material collection (initialized with msm btf) */
-    std::shared_ptr<mesh_gl::GPUMaterialCollection> m_material_collection;
-
-    uint32_t m_version;
-};
-
-} // namespace megamol::archvis_gl
